Added UDP Protocol

+++ Game Server/Constants.cs
@@ -3,6 +3,6 @@ namespace Game_Server
     public class Constants
     {
         public const int TICKS_PER_SEC = 30;
-        public const int MS_PER_TICK = 1000 / TICKS_PER_SEC;
+        public const float MS_PER_TICK = 1000f / TICKS_PER_SEC;
     }
 }
